Add AImageDecoder HeaderInfo_get/_setDataSpace
Bug: 135133301
Test: Iffe659e50078139188c3325545624640ae177cc2

Default to ADATASPACE_UNKNOWN for images that do not have a matching
ADataSpace. This matches the behavior of android.graphics.ImageDecoder
and BitmapFactory, which would use an "Unknown" ColorSpace for the same
images. It also means that a client who knows the true underlying
profile information to treat it as if it came from that profile.
Further, it means that if we add more ADataSpace values in the future,
the decode will be unchanged. A client can still choose to use a known
ADataSpace using _setDataSpace.

Change-Id: Id3ecf3bab17e4905fb2b75410ec756233d600c97