SF: DispSync should sync to activeConfig period
HWC@2.4 has the getDisplayVSyncPeriod method added, which was
being queried around rate-change events, and the returned value
was sent to the DispSync system as the period to synchronize to.

Some devices will have irregular vsync intervals coming out of certain
power-saving states, and the HWC was reporting an irregular vsync
period, relative to the activeConfig's period. This could manifest as
transient de-synchronization in DispSync (currently used code), and
in the VSyncReactor system (switched off) a bad rate would be adopted.

Fixes: 147734678
Test: uibench with VSyncReactor system
Test: uibench with DispSync system

Change-Id: Ia98b3f09893004b4420487385e0d1653dcfd2d8d