Clean up imagedecoder header file
Bug: 135133301
Test: Ibf7c0e563feeb08ce6dbabb5e86ddb385c9dff54

Remove AImageDecoderHeaderInfo_isAnimated. We are punting animation
support to S, so there is no reason to report whether the image is
animated.

Use int32_t for width and height. We already return these for
AImageDecoderHeaderInfo_getWidth/getHeight, so use the same type for
AImageDecoder_setTargetSize/computeSampledSize.

Use int32_t for AImageDecoderHeaderInfo_getAndroidBitmapFormat. This
matches the convention for what to return when the value is logically an
enum.

Change-Id: I93df851dd9fee2eb8d097e2158fb95003a0474db