Update imagedecoder.h to build as a C file
Bug: 135133301
Test: build as a C file

AAsset is a forward-declared struct, so it needs to be declared in
parameters as a struct in order to compile as a C file.

In addition, include rect.h explicitly for documentation purposes. This
also means that ARect as a parameter does not need to be declared as
"struct ARect".

Change-Id: I30c5176010ec3b5b0405f0b383654a7cfca65b62