Use cc_defaults in inputflinger targets
We add a cc_defaults for all of the targets in inputflinger, and include
those defaults in inputflinger_tests. This makes it so that building
the test compiles all of the sources in the test target and includes
them in the executable.

The primary benefit of this refactor is that running the test on any
device will now test the inputflinger code that the test was compiled
with instead of testing the libraries that were on the device. Also,
the device does not need to be re-flashed or need its libraries to be
updated when running the tests after modifying inputflinger code.

Bug: None
Test: atest inputflinger_tests
Change-Id: Ic0406355e5b485c8bf29a65f7c7c37ea8cb68eab