Remove device state when device is reset
When device reset is received in MotionClassifier, don't just set the
classification for that device to 'none'.
Instead, remove all state related to that device.

This prevents growing indefinitely the unordered_maps
that contain device-related state.

Also, log the feature state to the InputClassifier dump.

Bug: 150419367
Test: run several touch integration tests several times, and then do
'dumpsys input'. Ensure that the list doesn't grow as the tests are
executed multiple times.

Change-Id: I2e1ac359458a321f87e4599bb19350623afc9b2b