Remove frame rate functions from buffer queue
Remove the frame rate functions that were recently added to buffer queue
in ag/10028291. Modifying buffer queue is considered risky because we
have prebuilt binaries that use the classes in unknown ways.

We added an alternative mechanism for communicating frame rate info in
ag/10221036, so this code is no longer used for anything.

Bug: 150642189

Test: - Built and deployed to a Pixel 4, confirmed normal behavior.
- Built all the surface flinger unit tests.
- Ran the setFrameRate() cts test.

Change-Id: I26a620e842e79a378f58a78a088cff872a49de0d