Reduce setInputWindows calls
The compare function 'Region::isTriviallyEqual' is just compare the
first iterator between two regions, and it will always get false if we
passed it from binder that would create a new one even they are equal.

That would cause the layer to recompute the visible regions and make
it direct to call 'updateInputWindowInfo' continually.

- Change to use Region::hasSameRects to reduce some unnecessary calls.

Bug: 133780957
Test: manual, open debug logs and watch a video
Change-Id: Ibc9c488c179b403b203a1853cd6f3be9bb23fbaa