SF: Remove a redundant write
The Layer compositionState structure was only intended to be updated
during the call from the LayerFE::prepareCompositionState() call, which
in turn calls preparePerFrameCompositionState().

Both BufferQueueLayer and BufferStateLayer were updating the buffer
entry in the composition state structure in updateActiveBuffer(), which
was redundant with an update of the buffer entry also being made there.

This *may* also have resulted in the StrongPointer code reporting a
write race, resulting in a reported bug.

Test: go/wm-smoke
Test: atest CtsColorModeTestCases
Test: atest CtsDisplayTestCases
Test: atest CtsGraphicsTestCases
Test: atest CtsUiRenderingTestCases
Test: atest CtsViewTestCases
Test: atest android.media.cts.EncodeVirtualDisplayWithCompositionTest
Bug: 149663608
Change-Id: I5b2fc552a92746adfe3693059dab315a573a4860