SurfaceFlinger: Avoid overlapping calls to BOOT_FINISHED
Handling BOOT_FINISHED we write to mInputFlinger and mWindowManager.
This could trigger crashes if we write from multiple concurrent threads
or if we try and use it from the main thread while decing the strong
pointer. The easiest way to work around this seems to just to only
be willing to handle BOOT_FINISHED once, until the WM dies.

Bug: 150227563
Bug: 150225569
Test: Existing tests pass
Change-Id: I10f4b9359a5f3dc49e3360f28fbe808f0fc2afc8