SurfaceFlinger: Ignore mInputDirty if syncing input-windows.
It's possible we sync input windows but then don't actually have any
changes resulting in input being dirtied. This could lead to extremly
long waits on Binder threads and "Set transaction state timed-out"
errors. To fix this we ignore the cache state if we are processing a
sync command. I thought about just not invoking setInputWindows and
signalling the CV but I thought maybe someone could be relying on
the fencing semantics of syncInputWindows even in cases where
this transaction had no change.

Bug: 151182359
Test: android.media.cts.EncodeVirtualDisplayWithCompositionTest
Change-Id: Ia4e5f21c37db68a1a3bec36051c34df3be5fce52