Clean up some formatting problems in the setFrameRate ndk docs
Bug: 152413112

Test: Unfortunately there's no convenient way to build the ndk docs
locally, so we're just going to submit and see if the regenerated DAC
docs look correct. I talked with Dan Albert on the ndk team about this,
and he decided to reopen b/115644359 to see if we can get some easy way
to build the ndk docs locally again, so we can verify minor formatting
fixes like this.

Change-Id: Ie29a19a386887df6305ee9a4447702126fed7972