Reallocate blur textures if they don't fit
RenderEngine#drawLayers might be called with multiple display sizes.
Sometimes, the display might be quite small, not reflecting a physical
device. This happens becuse we support color sampling, and in those
cases we can render only a small portion of the output.

This could lead to FBO allocation that doesn't represent the surface
that we'll render to. To fix the issue, the textures need to be
resized when necessary.

Test: manual
Fixes: 152282559
Change-Id: I80d0c741b7f701a64fb862273433b167a97fe0bb